+++ src/main/java/pl/zagorski/services/MedicineImpl.java
@@ -8,10 +8,10 @@ import pl.zagorski.exceptions.ExceptionSample;
 import java.util.List;
 
 public interface MedicineImpl {
-    void save(int idPrescription, int idCharacter, int idProducer, String name, double price, double discount,
+    void save(int idPrescription, int idCharacter, int idProducer, String name, double price, int discount,
               String portion, String wrapping);
     void edit(int idMedicineEdit,int idPrescriptionEdit,int idCharacterEdit, int idProducerEdit,
-              String nameEdit,double priceEdit,double discountEdit, String portionEdit,String wrappingEdit);
+              String nameEdit,double priceEdit,int discountEdit, String portionEdit,String wrappingEdit);
     void delete(int idMedicineDelete) throws ExceptionSample;
     List<Medicine> findAll();
     Medicine findOne(int id);
